Issue #3891: reorganized TreeWalkerTest inputs
