--- src/test/resources/com/puppycrawl/tools/checkstyle/InputMain.java
@@ -1,6 +1,0 @@
-package com.puppycrawl.tools.checkstyle;
-/*comment*/
-public class InputMain {
-}
-class InputMainInner {
-}
