--- src/test/resources/com/puppycrawl/tools/checkstyle/import-control_complete.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE import-control PUBLIC
-    "-//Puppy Crawl//DTD Import Control 1.0//EN"
-    "http://checkstyle.sourceforge.net/dtds/import_control_1_0.dtd">
-
-<import-control pkg="com">
-  <allow class="some.class"/>
-  <disallow class="another.class" local-only="true"/>
-  <allow pkg="some.pkg"/>
-  <disallow pkg="another.pkg" local-only="true"/>
-  <disallow pkg="and.another.pkg" exact-match="true"/>
-  <subpackage name="puppycrawl">
-    <disallow pkg="some.pkg"/>
-    <disallow class="some.class"/>
-  </subpackage>
-</import-control>
