+++ src/test/resources/com/puppycrawl/tools/checkstyle/treewalker/InputTreeWalker.java
@@ -1,0 +1,6 @@
+package com.puppycrawl.tools.checkstyle.treewalker;
+/*comment*/
+public class InputTreeWalker {
+}
+class InputTreeWalkerInner {
+}
