+++ src/test/resources/com/puppycrawl/tools/checkstyle/treewalker/InputTreeWalkerHiddenComments.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle;
+package com.puppycrawl.tools.checkstyle.treewalker;
 
 /**
  * Some Javadoc.
